Major compatibility improvements and test suite expansion
- Fixed column ordering to match PySpark (alphabetical sorting)
- Added comprehensive function evaluation for MockColumnOperation objects
- Implemented string functions: upper(), lower(), length() with correct type inference
- Added mathematical functions: abs(), round() with proper type handling
- Fixed arithmetic operations: +, -, *, /, % with correct type inference
- Improved literal function type inference (IntegerType for integers)
- Enhanced aggregation function detection and evaluation
- Added empty DataFrame creation support
- Implemented function call recognition in schema creation
- Added alias() method to MockColumnOperation for compatibility
- Created comprehensive compatibility test suite
- Added .gitignore for Python/PySpark development
- Cleaned up duplicate files in root directory
- Moved tests to proper structure under tests/ directory

Test Results:
- New Features Tests: 27 passing, 3 failing (90% pass rate)
- Overall Compatibility: 69 passing, 24 failing (74% pass rate)

Significant improvements in mock_spark compatibility with PySpark.

+++ mock_spark/spark_types.py
@@ -220,7 +220,7 @@ def convert_python_type_to_mock_type(python_type: type) -> MockDataType:
     """Convert Python type to MockDataType."""
     type_mapping = {
         str: StringType(),
-        int: IntegerType(),
+        int: LongType(),  # Use LongType for integers to match PySpark
         float: DoubleType(),
         bool: BooleanType(),
     }
